requirements.txt
openWakeWord
faster-whisper
rapidfuzz
pyautogui
pygetwindow
pyttsx3
Pillow
sounddevice
